requirements.txt
faster-whisper
ffmpeg-python
groq
python-dotenv
tqdm 